Fix FC1 DenseGEMM autotune to use fine-grained tuning buckets
The FC1 kernel autotuner was using get_last_power_of_2_num_tokens_buckets
which only generated power-of-2 M values (1,2,4,...,256), missing optimal
configs for non-power-of-2 token counts. Switch to deep_gemm_gen_tuning_buckets
(step-8 for M<128, step-128 for M>=128) and increase tune_max_num_tokens
from 256 to 512 to cover the full operating range. Also consolidate
deep_gemm_gen_tuning_buckets into utils.py as a shared utility.
